pyproject.toml
[project]
name = "schedule-bot"
version = "0.1.0"
description = "A discord bot for scheduling TTRPG sessions with your group"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "hikari>=2.3.5",
    "python-dotenv>=1.1.1",
]
